requirements.txt
aws-cdk-lib>=2.232.2,<3.0.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_glue_alpha